% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh2md.r
\name{gh2md}
\alias{gh2md}
\title{Download a GitHub issue and its comments as Markdown#}
\usage{
gh2md(repo, issue, file)
}
\arguments{
\item{repo}{Repository as URL ("https://github.com/owner/repo/issues/52"),
shorthand ("owner/repo@52"), or "owner/repo" string.}

\item{issue}{Issue number. Not needed when encoded in `repo`.}

\item{file}{Optional file path to write the output to.}
}
\value{
Character string of Markdown (invisibly when `file` is given).
}
\description{
Download a GitHub issue and its comments as Markdown#
}
\examples{
\dontrun{
gh2md("https://github.com/JBGruber/atrrr/issues/52")
gh2md("JBGruber/atrrr@52")
gh2md("JBGruber/atrrr", 52)
gh2md("JBGruber/atrrr@52", file = "52.md")
}

}
