% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_deps.r
\name{install_deps}
\alias{install_deps}
\title{Install dependencies}
\usage{
install_deps(path = ".", packages = NULL, gh_packages = NULL)
}
\arguments{
\item{path}{path to check for files (checks recursivly).}

\item{packages}{additional packages.}

\item{gh_packages}{additional GitHub packages.}
}
\description{
Checks the path for R, Rmd and Qmd files and installs all references packages.
}
\examples{
\dontrun{
install_deps(gh_packages = "JBGruber/paperboy")
}
}
