% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.r
\name{make_quarto}
\alias{make_quarto}
\title{make a project containing quarto document}
\usage{
make_quarto(files, destfiles = NULL, output_ext = ".html", ...)
}
\arguments{
\item{files}{the quarto source files.}

\item{destfiles}{optional destination files. Names are created from input and
output_ext if NULL.}

\item{output_ext}{alternatively to destfiles you can supply an output
extension}

\item{...}{passed on to \link[quarto]{quarto_render}.}
}
\description{
A Make-like pipeline function that checks if source files have been updated
and renders them if the html output is older than the source file.
\code{files} are made in the order they are supplied and the function checks
if each file's output is newer than the last and whether the output is newer
than the source file.
}
\examples{
\dontrun{
make_quarto(list.files(ist.files(pattern = "\\\\.qmd$")))
}
}
