% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_template.r
\name{report_template}
\alias{report_template}
\title{Create a new Quarto report file}
\usage{
report_template(name = NULL)
}
\arguments{
\item{name}{optional name}
}
\description{
Create a new Quarto report file
}
